인터페이스 분리 원칙 적용
인터페이스 분리 원칙 적용
